+++ rosomaxa/src/evolution/config.rs
@@ -305,28 +305,29 @@ where
     pub fn build(self) -> Result<EvolutionConfig<C, O, S>, String> {
         let context = self.context.ok_or_else(|| "missing heuristic context".to_string())?;
         let logger = context.environment().logger.clone();
-        let heuristic = if let Some(heuristic) = self.heuristic {
-            (logger)("configured to use custom heuristic");
-            heuristic
-        } else {
-            Box::new(DynamicSelective::new(
-                self.search_operators.ok_or_else(|| "missing search operators or heuristic".to_string())?,
-                self.diversify_operators.ok_or_else(|| "missing diversify operators or heuristic".to_string())?,
-                context.environment(),
-            ))
-        };
         let termination =
             Self::get_termination(&logger, self.max_generations, self.max_time, self.min_cv, self.target_proximity)?;
 
         Ok(EvolutionConfig {
             initial: self.initial,
-            context,
             strategy: if let Some(strategy) = self.strategy {
                 (logger)("configured to use a custom strategy");
                 strategy
             } else {
+                let heuristic = if let Some(heuristic) = self.heuristic {
+                    (logger)("configured to use custom heuristic");
+                    heuristic
+                } else {
+                    Box::new(DynamicSelective::new(
+                        self.search_operators.ok_or_else(|| "missing search operators or heuristic".to_string())?,
+                        self.diversify_operators
+                            .ok_or_else(|| "missing diversify operators or heuristic".to_string())?,
+                        context.environment(),
+                    ))
+                };
                 Box::new(strategies::Iterative::new(heuristic, 1))
             },
+            context,
             termination,
             processing: self.processing,
         })